crust owns its Android build contract via its own config plugin
Every host embedding crust needed three hand-mirrored gradle edits: the
authenticated Mapbox Downloads repo (the Nav SDK's artifacts exist only
in Mapbox's private registry), a global protobuf-javalite exclusion
(duplicate classes against the stack's protobuf-java), and core-library
desugaring (crust's AAR metadata demands it). The Mentra app carried
them in plugins/android.ts; the example OEM app carried a copy in
withMentraStackGradle; a real OEM would have written a third.

crust now ships an expo config plugin (the bluetooth-sdk pattern:
app.plugin.js -> plugin/build) carrying all three — with the hardened
placement and per-injection idempotency markers from the example app's
copy. Both hosts list "@mentra/crust" in their plugins and both
hand-mirrored copies are deleted. Verified by fresh prebuilds of both
apps: all three injections present in the generated gradle files.

Follow-up decision (PR body): making the Nav dependency opt-in, which
would drop the MAPBOX_DOWNLOADS_TOKEN requirement for hosts that don't
navigate.
