+++ mobile/modules/crust/app.plugin.js
@@ -1,0 +1,3 @@
+const plugin = require("./plugin/build")
+
+module.exports = plugin.default ?? plugin
